use the genericname as comment
In absence of an appdata file, appstream tries to determine its required
fields through the desktop file, currently ksquares is missing a summary
which would get pulled from the Comment entry in the desktop file.
Since the GenericName can easily double as (not all too grand) Comment
here, simply copy it to reduce l10n work.

Not having appstream data prevents the application from showing up
in software stores we'll want this fixed in a bugfix release of 16.12.

REVIEW: 129888